requirement.txt
ipykernel
numpy
pandas
matplotlib
seaborn
flask
requests
streamlit
scikit-learn
imblearn
openpyxl
statsmodels
pipelining
xgboost
lxml
html5lib
kneed
dill
catboost
PyYAML
#-e .